backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<1.1
django-cors-headers>=4.0,<4.4
dj-rest-auth>=5.0,<5.1 # Added for easier auth endpoint setup
djangorestframework-simplejwt>=5.3,<5.4 # Needed by dj-rest-auth for JWT
